+++ plugin-animation/api/plugin-animation.api
@@ -122,3 +122,11 @@ public final class com/mapbox/maps/plugin/animation/animator/Evaluators {
 	public final fun getSCREEN_COORDINATE ()Landroid/animation/TypeEvaluator;
 }
 
+public final class com/mapbox/maps/plugin/animation/threads/MapboxAnimatorThread {
+	public fun <init> ()V
+	public final fun onAttached ()V
+	public final fun onDetached ()V
+	public final fun post (Ljava/lang/Runnable;)Z
+	public final fun postMain (Ljava/lang/Runnable;)Z
+}
+
