+++ plugin-animation/src/test/java/com/mapbox/maps/plugin/animation/threads/MapboxAnimatorThreadTest.kt
@@ -1,4 +1,4 @@
-package com.mapbox.maps.plugin.animation
+package com.mapbox.maps.plugin.animation.threads
 
 import android.os.Build
 import android.os.Looper
