+++ sdk-base/api/sdk-base.api
@@ -888,14 +888,6 @@ public final class com/mapbox/maps/plugin/animation/MapAnimationOwnerRegistry {
 	public static final field LOCATION Ljava/lang/String;
 }
 
-public final class com/mapbox/maps/plugin/animation/MapboxAnimatorThread {
-	public fun <init> ()V
-	public final fun onAttached ()V
-	public final fun onDetached ()V
-	public final fun post (Ljava/lang/Runnable;)Z
-	public final fun postMain (Ljava/lang/Runnable;)Z
-}
-
 public abstract class com/mapbox/maps/plugin/annotation/Annotation {
 	public static final field Companion Lcom/mapbox/maps/plugin/annotation/Annotation$Companion;
 	public static final field ID_DATA Ljava/lang/String;
